requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.44
botocore==1.34.44
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
crispy-tailwind==1.0.3
cryptography==42.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-heroku==0.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
feedparser==6.0.11
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
whitenoise==6.6.0
